requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
chardet==5.2.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.0.0
python-bidi==0.6.3
reportlab==4.2.5
sqlparse==0.5.1
